LPD-26139 Add type prop that is missing on Select component from ObjectField
With this we fix a bug where the label wasn't recieving the id prop, thus not being able to connect the field input with its label. This happened because there is a condition that checks if the field type is 'select' to add the id prop, but it was coming with no value before.