LPD-37753 Ignore root update trigger option when the object inheritance is disabled
If the inheritance is disabled after creating an action related to the root update, the action will still be there but with the trigger option not selected. If the inheritance is enabled again, the action will work as expected with the trigger of root update.

Notice that the trigger can be changed to other options when the inheritance is disabled as well.